# src/AppBundle/CSPro/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Web API for synchronizing data and files from devices in the field to a web server.\n",
    "version": "2.0.0",
    "title": "CSPro Sync",
    "contact": {
      "name": "cspro@lists.census.gov"
    },
    "license": {
      "name": "Public domain"
    }
  },
  "host": "csweb.teleyah.com",
  "basePath": "/api/",
  "schemes": [
    "http",
    "https"
  ],
  "produces": [
    "application/json",
    "text/plain",
    "application/zip"
  ],
  "paths": {
...
    "/files/{filePath}": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "Get file",
        "description": "Get contents of file matching path.",
        "operationId": "GetFileInfo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "filePath",
            "description": "path of file to download for",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FileInfo"
            }
          },
          "403": {
            "description": "Unable to read file"
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    },
    "/files/{filePath}/content": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "Get file",
        "description": "Get contents of file matching path.",
        "operationId": "GetFileContent",
        "produces": [
          "application/octet-stream"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "filePath",
            "description": "path of file to download",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "If-None-Match",
            "type": "string",
            "description": "ETag value from previous response, 304 is returned if this matches current ETag"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "file"
            },
            "headers": {
              "Content-MD5": {
                "description": "Signature of file contents",
                "type": "string"
              },
              "Content-Length": {
                "description": "Size in bytes of downloaded file",
                "type": "integer",
                "format": "int64"
              },
              "ETag": {
                "description": "Signature of file for caching",
                "type": "string"
              }
            }
          },
          "304": {
            "description": "file not modified since last time downloaded according to ETag"
          },
          "403": {
            "description": "Unable to read file"
          },
          "404": {
            "description": "File not found"
          }
        }
      },
      "put": {
        "tags": [
          "file"
        ],
        "summary": "Upload file",
        "description": "Put contents of file to server directory specified by filePath. Create directory filePath if it does not exist.",
        "operationId": "PutFileContent",
        "consumes": [
          "application/octect-stream"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "filePath",
            "description": "Full path to upload file to",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "Content-MD5",
            "required": true,
            "description": "Signature of file contents",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Response is metadata for uploaded file.",
            "schema": {
              "$ref": "#/definitions/FileInfo"
            }
          },
          "403": {
            "description": "Unable to write to filePath"
          },
          "500": {
            "description": "Error writing file"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "FileInfo": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "description": "either \"file\" or \"directory\""
        },
        "name": {
          "type": "string",
          "description": "name of file or directory without path"
        },
        "directory": {
          "type": "string",
          "description": "absolute of path directory containing file (for files) or absolute path of parent directory (for directories)"
        },
        "md5": {
          "type": "string",
          "description": "md5 signature of file (file only)"
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "description": "size of file in bytes (file only)"
        },
        "lastModified": {
          "type": "string",
          "format": "date-time",
          "description": "date/time of last modification of file in RFC3339 format ( e.g. 1985-04-12T23:20:50.52Z)"
        }
      }
    },
...
  }
}